First working version with console only
Objective is to have a layered approach allowing to use rp_core totally
independantly from the B2SAFE underlying technology.
rp_console can be launched to test based and uses text files as inputs
and outputs
The B2SAFE ruleset rp_ingest is the one created for CUNI replication. It
will be updated in the future to avoid any duplication of B2SAFE core
functions.

This is using a temporary version of ReplicationService class provided
by CUNI